feat(r-type): leaderboard complete

+++ client/ui/elements/focusable/Dropdown.cpp
@@ -302,16 +302,16 @@ math::Vector2f Dropdown::getOptionPosition(size_t optionIndex) const {
     switch (_direction) {
         case DropdownDirection::Down:
             return {basePos.getX(), basePos.getY() + size.getY() +
-                OPTION_SPACING + (size.getY() + OPTION_SPACING) * optionIndex};
+                OPTION_SPACING + (size.getY() + OPTION_SPACING) * static_cast<float>(optionIndex)};
         case DropdownDirection::Up:
             return {basePos.getX(), basePos.getY() - size.getY() -
-                OPTION_SPACING - (size.getY() + OPTION_SPACING) * optionIndex};
+                OPTION_SPACING - (size.getY() + OPTION_SPACING) * static_cast<float>(optionIndex)};
         case DropdownDirection::Right:
             return {basePos.getX() + size.getX() + OPTION_SPACING +
-                (size.getX() + OPTION_SPACING) * optionIndex, basePos.getY()};
+                (size.getX() + OPTION_SPACING) * static_cast<float>(optionIndex), basePos.getY()};
         case DropdownDirection::Left:
             return {basePos.getX() - size.getX() - OPTION_SPACING -
-                (size.getX() + OPTION_SPACING) * optionIndex, basePos.getY()};
+                (size.getX() + OPTION_SPACING) * static_cast<float>(optionIndex), basePos.getY()};
         default:
             return basePos;
     }
@@ -326,11 +326,11 @@ math::Vector2f Dropdown::getDropdownSize() const {
     switch (_direction) {
         case DropdownDirection::Down:
         case DropdownDirection::Up:
-            size.setY(size.getY() * (optionCount + 1));
+            size.setY(size.getY() * static_cast<float>(optionCount + 1));
             break;
         case DropdownDirection::Left:
         case DropdownDirection::Right:
-            size.setX(size.getX() * (optionCount + 1));
+            size.setX(size.getX() * static_cast<float>(optionCount + 1));
             break;
     }
